Cast to float.

+++ src/compositor/seat/pointer.c
@@ -158,8 +158,8 @@ wlc_pointer_focus(struct wlc_pointer *pointer, struct wlc_view *view, struct wlc
       if (!(s = convert_from_wlc_resource(view->surface, "surface")))
          return;
 
-      d.x = (pointer->pos.x - v.origin.x) * s->size.w / v.size.w;
-      d.y = (pointer->pos.y - v.origin.y) * s->size.h / v.size.h;
+      d.x = (pointer->pos.x - v.origin.x) * (float)s->size.w / v.size.w;
+      d.y = (pointer->pos.y - v.origin.y) * (float)s->size.h / v.size.h;
 
       if (out_pos)
          memcpy(out_pos, &d, sizeof(d));
